Backend-form/src/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
arrow==1.3.0
asgiref==3.7.2
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.3
billiard==4.2.0
blessed==1.20.0
boltons==23.1.1
boto3==1.34.52
botocore==1.34.52
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
config-client==1.4.0
coreapi==2.3.3
coreschema==0.0.4
coverage==7.3.4
croniter==2.0.2
cryptography==39.0.2
Django==4.1
django-cors-headers==4.3.1
django-debug-toolbar==4.2.0
django-easy-audit==1.3.5
django-extensions==3.2.3
django-filter==23.5
django-jazzmin==2.6.0
django-picklefield==3.1
django-pytest==0.2.0
django-q==1.3.9
django-redis==5.4.0
django-rest-passwordreset==1.4.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
dnspython==2.4.2
docopt==0.6.2
drf-spectacular==0.27.1
ecdsa==0.19.0
exceptiongroup==1.2.0
executing==2.0.1
face==20.1.1
Faker==12.0.1
fido2==1.0.0
glom==23.5.0
icecream==2.1.3
idna==3.6
ifaddr==0.2.0
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.5
linecache2==1.0.0
MarkupSafe==2.1.3
mixer==7.2.2
mysql-connector-python==8.3.0
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==23.2
pipreqs==0.4.13
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
py-eureka-client==0.11.8
pyasn1==0.6.0
pycparser==2.22
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
PyJWT==2.8.0
PyMySQL==1.1.0
pyotp==2.9.0
pypng==0.20220715.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-django==4.7.0
pytest-json-report==1.5.0
pytest-metadata==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
python-jose==3.3.0
python-u2flib-server==5.0.1
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
referencing==0.33.0
requests==2.31.0
rpds-py==0.18.0
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3transfer==0.10.0
simplejson==3.19.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==8.4.0
termcolor==2.4.0
tomli==2.0.1
traceback2==1.4.0
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.3
ua-parser==0.18.0
unittest2==1.1.0
uritemplate==4.1.1
urllib3==2.0.7
user-agents==2.2.0
vine==5.1.0
wcwidth==0.2.13
yarg==0.1.9
